refactor: rewire all type importers to new locations; delete shim (#632 segment 2 / A2)
Step A2 of types split: rewrite every importer to point at the new
per-layer types files; delete the legacy/core/types.ts re-export shim.

Symbol → destination map (final):
- Tool, AgentToolSettings           → src/tools/types.js
- ProviderConfig, AgentConfig,
  CompactionMode, CompactionConfig  → src/agent/types.js
- Session, SessionMetadata,
  SessionConfig, SessionStoreConfig,
  SessionStore                      → src/sessions/types.js
- PeerKind, BindingPeer,
  BindingMatch, Binding,
  ChannelsConfig, Transport         → src/gateway/types.js
- CronActionConfig                  → src/automation/types.js
- AgentMessage, AgentEvent          → @mariozechner/pi-agent-core (direct)
- Usage, ImageContent               → @mariozechner/pi-ai (direct)

Deleted:
- src/legacy/core/types.ts (shim)
- src/legacy/core/types.test.ts (tested the shim's structure; no longer relevant)

Touched importers (~28 files):
- src/config.ts (split 11 symbols across 5 destinations)
- src/legacy/{tools,plugins,workspace,agents,commands,core}/...
- src/legacy/plugins/discord/{discord,discord-manager,index}.ts
- src/legacy/plugins/http/sessions.ts
- src/legacy/core/runtime.ts (inline import("./types.js") changed)
- src/gateway/bindings.{ts,test.ts} (transitional ../legacy/core path replaced with sibling ./types.js)

tsc clean; 1158/1158 tests pass (= 1165 minus 7 shim-structure tests).
