+++ src/legacy/commands/slash-commands.ts
@@ -1,7 +1,7 @@
 // src/commands/slash-commands.ts — Slash command handler for admin operations
 // Parses and dispatches /status, /reload, /model commands from chat messages.
 
-import type { SessionStore } from "../core/types.js";
+import type { SessionStore } from "../../sessions/types.js";
 import type { AgentServiceCache } from "../core/pi-mono.js";
 import type { DefaultAgentManager } from "../core/agent-manager.js";
 import { createLogger } from "../../logging/logger.js";
