+++ src/legacy/plugins/discord/discord-manager.ts
@@ -3,7 +3,7 @@
 
 import type {
   SessionStore,
-} from "../../core/types.js";
+} from "../../../sessions/types.js";
 import type { DiscordAccountConfig } from "./types.js";
 import type { DefaultAgentManager } from "../../core/agent-manager.js";
 import type { AgentRuntime } from "../../agents/runtime.js";
