+++ src/legacy/plugins/discord/discord.test.ts
@@ -2,7 +2,7 @@
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { DiscordTransport } from "./discord.js";
-import type { SessionStore } from "../../core/types.js";
+import type { SessionStore } from "../../../sessions/types.js";
 import type { AgentServiceCache } from "../../core/pi-mono.js";
 import type { DefaultAgentManager } from "../../core/agent-manager.js";
 import { ThreadBindingManager } from "./thread-bindings.js";
